Refactored to tidy up files

+++ mySQL/functions.php
@@ -3,7 +3,35 @@ include "db.php";
 
 // a series of functions we are using to query our database
 
+function createRecord()
+{
+    global $connection;
+    if (isset($_POST['submit'])) {
+        $username = $_POST['username'];
+        $password = $_POST['password'];
 
+
+        //writing a query to a database:
+        $query = "INSERT INTO users(username,password)";
+
+        //concatenate our query:
+        $query .= "VALUES ('$username', '$password')";
+
+        //the above is a neater way to submitting the query rather than putting all of lines 15-22 inside of mysqli_query()
+
+        $result = mysqli_query($connection, $query);
+
+
+
+        if (!$result) {
+            die('Query Failed ');
+        } else {
+            echo "Record Created!";
+        }
+    }
+}
+
+//Retrieve
 function showAllData()
 {
     global $connection;
@@ -24,7 +52,6 @@ function showAllData()
         echo "<option value='$id'>$id</option>";
     }
 }
-//CRUD:
 
 //Update: 
 
@@ -32,37 +59,47 @@ function showAllData()
 
 function updateTable()
 {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if (isset($_POST['submit'])) {
+        global $connection;
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $id = $_POST['id'];
 
-    $query = "UPDATE users SET ";
-    $query .= "username = '$username', ";
-    $query .= "password = '$password' ";
-    $query .= "WHERE id = $id";
+        $query = "UPDATE users SET ";
+        $query .= "username = '$username', ";
+        $query .= "password = '$password' ";
+        $query .= "WHERE id = $id";
 
 
-    $result = mysqli_query($connection, $query);
-    if (!$result) {
-        die("QUERY FAILED " . mysqli_error($connection));
+        $result = mysqli_query($connection, $query);
+        if (!$result) {
+            die("QUERY FAILED " . mysqli_error($connection));
+        } else {
+            echo "Record Updated";
+        }
     }
 }
 
+//Delete
+
 function deleteRows()
 {
-    global $connection;
-    $username = $_POST['username'];
-    $password = $_POST['password'];
-    $id = $_POST['id'];
+    if (isset($_POST['submit'])) {
+        global $connection;
+        $username = $_POST['username'];
+        $password = $_POST['password'];
+        $id = $_POST['id'];
 
-    $query = "DELETE FROM users ";
+        $query = "DELETE FROM users ";
 
-    $query .= "WHERE id = $id";
+        $query .= "WHERE id = $id";
 
 
-    $result = mysqli_query($connection, $query);
-    if (!$result) {
-        die("QUERY FAILED " . mysqli_error($connection));
+        $result = mysqli_query($connection, $query);
+        if (!$result) {
+            die("QUERY FAILED " . mysqli_error($connection));
+        } else {
+            echo "Record Deleted";
+        }
     }
 }
